+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.4.3-SNAPSHOT
+- [java-security] Supports Validation for SAP ID Tokens which results into these incompatible changes:
+  - constant `TokenClaims.XSUAA.EMAIL` -> `TokenClaims.EMAIL`
+  - constant `TokenClaims.XSUAA.USER_NAME` -> `TokenClaims.USER_NAME`
+  - constant `TokenClaims.XSUAA.FAMILY_NAME` -> `TokenClaims.FAMILY_NAME`
+  - constant `TokenClaims.XSUAA.GIVEN_NAME` -> `TokenClaims.GIVEN_NAME
+  - api rename Token.getAccessToken()` renamed to `Token.getTokenValue()`
+  - api deleted `Token.getBearerAccessToken()` as it fits not to a ID token. You can compensate with `"Bearer " + Token.getTokenValue()`
+  - api move `Token.getGrantType()` to `AccessToken.getGrantType()` interface.
+- [java-security] `XSUserInfoAdapter` supports `isInForeignMode()` method.
+  
+  
 ## 2.4.2-SNAPSHOT
 - [java-security] XSA Support.